Refactor: standardize adapter init with AdapterFactory & reusable DiffUtils
- Created `BaseAdapterFactory` interface with `DefaultBaseAdapterFactory` in `base/`
- Standardized `ResourcesAdapter` and `CoursesAdapter` with factory instantiation methods.
- Extracted generic `standardItemCallback` logic to `DiffUtils.kt` to minimize boilerplate.
- Updated `BaseRecyclerFragment` to optionally expose the `adapterFactory` variable.
